Add support for bashftp (https://github.com/alzwded/bashftp) and md5 checksums (but only in bashftp for now)

+++ vm/run_tests.sh
@@ -43,6 +43,10 @@ run_single_test() {
     echo $t
     echo '===================================================='
     for conf in confs/* ; do
+        if [[ -f ${t%.sh}.${conf##*/}.skip ]] ; then
+            echo Skipping $t for $conf
+            continue
+        fi
         echo $conf
         echo '----------------------------------------------------'
         rm -rf scratch/
